pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asr-project"
version = "0.1.0"
description = "ASR Project with multiple model implementations"
requires-python = ">=3.8"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "deprecation==2.1.0",
    "fastapi==0.115.12",
    "frozenlist==1.6.0",
    "google-api-core==2.24.2",
    "google-auth==2.40.1",
    "google-cloud-core==2.4.3",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "gotrue==2.12.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "librosa==0.10.2",
    "multidict==6.4.3",
    "numpy==2.2.6",
    "packaging==25.0",
    "pluggy==1.5.0",
    "postgrest==1.0.1",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==6.30.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "pytest==8.3.5",
    "pytest-cov==4.1.0",
    "pytest-mock==3.14.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "realtime==2.4.3",
    "requests==2.32.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.12.1",
    "starlette==0.46.2",
    "storage3==0.11.3",
    "strenum==0.4.15",
    "supabase==2.15.1",
    "supafunc==0.9.4",
    "typing-extensions>=4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "websockets>=11,<15",
    "yarl==1.20.0"
]

[project.optional-dependencies]
asr-utils = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "tqdm>=4.65.0",
    "soundfile>=0.12.1"
]

whisper = [
    "openai-whisper==20240930",
    "numpy>=2.2.6",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "tiktoken==0.9.0",
    "regex==2024.11.6",
    "more-itertools==10.7.0"
]

wav2vec = [
    "transformers==4.52.3",
    "huggingface-hub==0.32.0",
    "tokenizers==0.21.1",
    "safetensors==0.5.3",
    "hf-xet==1.1.2",
    "numpy>=2.2.6",
    "torch>=2.7.0",
    "torchaudio>=2.7.0"
]

mesolitica = [
    "transformers==4.52.3",
    "huggingface-hub==0.32.0",
    "tokenizers==0.21.1",
    "safetensors==0.5.3",
    "hf-xet==1.1.2",
    "numpy>=2.2.6",
    "torch>=2.7.0",
    "torchaudio>=2.7.0"
]

moonshine = [
    "useful-moonshine @ git+https://github.com/usefulsensors/moonshine.git@2f6282347950c8d711bd7319287babb87ec5a92d",
    "numpy>=2.2.6",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "transformers>=4.30.0",
    "efficientnet-pytorch==0.7.1",
    "einops==0.8.0",
    "segmentation-models-pytorch==0.3.2",
    "timm==0.6.12",
    "pretrainedmodels==0.7.4",
    "playsound3>=2.5.2",
    "pydub>=0.25.1"
]

vosk = [
    "vosk>=0.3.45",
    "numpy>=2.2.6",
    "tqdm>=4.65.0",
    "playsound3>=3.2.3",
    "soundfile>=0.12.1",
    "websockets>=11,<15"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true
